// ui/src/types/generated/LimitDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LimitTargetDto } from "./LimitTargetDto";

export type LimitDto = { id: number, target: LimitTargetDto, default_minutes: number, weekday_minutes: [number | null, number | null, number | null, number | null, number | null, number | null, number | null], enabled: boolean, timer_expires_utc: string | null, };